fix: Python 3 correctness and PEP 8 compliance improvements

+++ SU2_PY/topology_optimization.py
@@ -120,11 +120,9 @@ class Driver:
         # write inputs
         self._write_input(x)
 
-        # clear previous output and run direct solver
-        try:
+        # clear previous output if present and run direct solver
+        if os.path.exists(self._objValFile):
             os.remove(self._objValFile)
-        except OSError:
-            pass  # Ignore error if file does not exist
 
         try:
             sp.call(self._objValCommand, shell=True)
@@ -147,11 +145,9 @@ class Driver:
     def obj_der(self, x):
         # inputs written in obj_val_driver
 
-        # clear previous output and run direct solver
-        try:
+        # clear previous output if present and run direct solver
+        if os.path.exists(self._objDerFile):
             os.remove(self._objDerFile)
-        except OSError:
-            pass  # Ignore error if file does not exist
         N = x.shape[0]
         y = np.ndarray((N,))
 
@@ -197,11 +193,9 @@ class Driver:
     def con_der(self, x):
         # inputs written in obj_val_driver
 
-        # clear previous output and run solver
-        try:
+        # clear previous output if present and run solver
+        if os.path.exists(self._conDerFile):
             os.remove(self._conDerFile)
-        except OSError:
-            pass  # Ignore error if file does not exist
         N = x.shape[0]
         y = np.ndarray((N,))
